pyproject.toml
[project]
name = "graph-rag"
version = "0.1.0"
description = "Multi-Agent Real Estate Asset Manager with LangGraph"
requires-python = ">=3.12,<3.13"
dependencies = [
    "langgraph>=0.1.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.1",
    "polars>=1.0",
    "pyarrow>=14.0",
    "mlflow>=2.10",
    "streamlit>=1.30",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "tenacity>=8.2",
    "thefuzz>=0.22",
    "httpx>=0.27",
    "openai>=1.12",
    "cachetools>=5.3",
    "plotly>=5.18",
    "numpy>=1.26",
    "scikit-learn>=1.4",
]

[project.optional-dependencies]
dev = [
    "anyio>=4.13.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=4.1",
    "hypothesis>=6.92",
    "ruff>=0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "anyio>=4.13.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
